Some refactoring and return bad request instead of user not found when authenticating
