+++ package.json
@@ -36,6 +36,7 @@
     "@types/supertest": "^2.0.12",
     "husky": "^8.0.0",
     "jest": "^28.1.1",
+    "prettier": "^3.0.3",
     "prisma": "^5.2.0",
     "supertest": "^6.3.3",
     "ts-node": "^10.8.1",
