+++ src/core/errors/index.ts
@@ -1,5 +1,9 @@
-import { ValidationError } from './validation.error';
+import ValidationError from './validation.error';
+import NotFoundError from "./not-found.error";
+import RequestError from "./request.error";
 
 export {
   ValidationError,
+  NotFoundError,
+  RequestError,
 };
