+++ src/core/errors/not-found.error.ts
@@ -1,7 +1,7 @@
 import { RuntimeError } from './custom.error';
 import { ErrorType } from './error.type';
 
-export class NotFoundError extends RuntimeError {
+export default class NotFoundError extends RuntimeError {
  constructor(message: string) {
    super(message, ErrorType.NOT_FOUND_ERROR);
  }
