+++ src/core/errors/request.error.ts
@@ -1,7 +1,7 @@
 import { RuntimeError } from './custom.error';
 import { ErrorType } from './error.type';
 
-export class RequestError extends RuntimeError {
+export default class RequestError extends RuntimeError {
   constructor(message: string) {
     super(message, ErrorType.INVALID_REQUEST);
   }
