+++ src/core/errors/validation.error.ts
@@ -1,7 +1,7 @@
 import { RuntimeError } from './custom.error';
 import { ErrorType } from './error.type';
 
-export class ValidationError extends RuntimeError {
+export default class ValidationError extends RuntimeError {
   constructor(message: string) {
     super(message, ErrorType.VALIDATION_ERROR);
   }
