+++ src/core/usecases/delete-credit-card.usecase.ts
@@ -1,7 +1,7 @@
-import { CreditCardRepository } from '../repositories';
-import { log } from '../logger/logger';
-import { CreditCard } from '../entities';
-import { RequestError } from '../errors/request.error';
+import {CreditCardRepository} from '../repositories';
+import {log} from '../logger/logger';
+import {CreditCard} from '../entities';
+import {NotFoundError} from "../errors";
 
 const deleteCreditCardUsecase = async (id: string, repository: CreditCardRepository) => {
   log(`[deleteCreditCardUseCase]: getting credit card information id `, id);
@@ -9,7 +9,7 @@ const deleteCreditCardUsecase = async (id: string, repository: CreditCardReposit
 
   if (!creditCard) {
     log(`[deleteCreditCardUseCase]: getting credit card information id `, id);
-    throw new RequestError(`credit card not found for id ${id}`);
+    throw new NotFoundError(`credit card not found for id ${id}`);
   }
 
   log(`[deleteCreditCardUseCase]: deletting credit card information id `, id);
